modules/illustrator-app/services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.1",
    "debugpy>=1.8.16",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-dramatiq>=0.14.0",
    "django-environ>=0.12.0",
    "django-ninja>=1.4.3",
    "dramatiq-abort[redis]>=1.2.1",
    "dramatiq[redis,watch]>=1.18.0",
    "pymupdf>=1.26.4",
    "redis>=5.3.1",
    "regex>=2025.8.29",
    "semantic-text-splitter>=0.27.0",
    "torch>=2.8.0",
    "transformers>=4.56.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.16",
]
